requirements.txt
Flask>=3.0.0
SQLAlchemy>=2.0.0
Flask-Migrate>=4.0.0
Pydantic>=2.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
huggingface_hub>=0.20.0
